Watch all namespaces
Make the operator watch all namespaces but only allow reconciliation of
resources in the namespace where it is running. This is in preparation
to allow watching deletion of CatalogSources in target namespaces other
than the one the operator is running in.

Now that the operator watches all namespaces, watch and list roles
for the child resources have to be cluster roles given the way the
cached client works.